requirements.txt
opencv-python~=4.7.0.72
numpy~=1.23.5
matplotlib~=3.7.1
Pillow~=9.5.0
keras~=2.12.0